Fix ncurses library detection
The block removed here only decided whether to pass -lncurses to gcc or
not.

When LIBPATH here is not in sync with gcc's search-dirs -- LIBRARY_PATH
(without LD_) and -L options -- the decision is not necessarily correct.
In my case -lncurses was not passed to gcc and the linking failed even
though gcc would have been able to find the library.

It's safer to always have the switch for the programs that need ncurses.