Replace adventure with the BDS C version.
I tried to find the newest available version of adventure and this
one appears to be the only usable one. This one was downloaded from
http://rickadams.org/adventure/e_downloads.html, BDS C section,
and is the original 350 point version, ported to QNX sometime in the
1990s. The winning features are that it is in C and it does not have
that horrifying spaghetti code switch statement main that I have given
up on untangling. Being the original version is also a plus as that is
what bsdgames had before, so there should be no functional changes here.

The more developed versions are mostly in Fortran, with the few
C versions being worthy of IOCCC submission. Some also use the
special interpreter, adding a lot of complexity. The BDS version
may not be the most recent, but it looks the cleanest. So here it is.

Anyway, the biggest change from the original advqnx.tar.gz contents
was to embed the text databases and use named constants for all the
magic numbers. The rest was just cleaning up the code to fit in with
the rest of the games here.