cribbage: explicitly check for go on over 31
The interface prevents the player from selecting a card that would put
the count over 31, but if the AI routine can do so if it has no other
cards to play. If the human has a card to play but the AI does not, the
AI could end up playing a card going over 31. Fixes #8.