cribbage: stop AI from leading with a 5
Because 10-value cards are much more common than others it is important
to avoid playing score values 10 less than the scoring values 15, 30,
and 31. For this reason, leading with a 5 or playing a 2 when the count
is 18 is undesirable. The AI will now consider this when multiple
non-scoring plays can be made and will avoid the above scenario.